requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.40.0
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.25.2



